fix(OTPField): create helper file
